// src/app/(dashboard)/kho/types/warehouse.ts
export interface Warehouse {
  MaKho: string;
  TenKho: string;
  DiaChi: string;
  HinhAnh: string;
  GhiChu: string;
  ThuKho: string;
  [key: string]: any;
}

export interface WarehouseFormData {
  MaKho: string;
  TenKho: string;
  DiaChi: string;
  HinhAnh: string;
  GhiChu: string;
  ThuKho: string;
}